+++ visualization/app/codeCharta/state/store/appSettings/isLoadingMap/isLoadingMap.reducer.ts
@@ -1,0 +1,10 @@
+import { IsLoadingMapAction, IsLoadingMapActions, setIsLoadingMap } from "./isLoadingMap.actions"
+
+export function isLoadingMap(state: boolean = setIsLoadingMap().payload, action: IsLoadingMapAction): boolean {
+	switch (action.type) {
+		case IsLoadingMapActions.SET_IS_LOADING_MAP:
+			return action.payload
+		default:
+			return state
+	}
+}