+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 ### Added 🚀
 
 - Support for camel and kebab-case for ccsh arguments #772
+- RawTextParser for analysis #660
+- IndentationLevel as metric for RawTextParser #660
+- Show additional Pairing Rate of Selected Building, simultaneously to the currently hovered Buildings #736
 
 ### Changed